🤖 refactor: 이벤트 상세페이지 서버사이드 prefetchQuery 적용
